refactor using extension functions
