+++ feature/player/src/main/java/dev/anilbeesetti/nextplayer/feature/player/PlayerActivity.kt
@@ -19,8 +19,6 @@ import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import androidx.media3.common.C
 import androidx.media3.common.MediaItem
-import androidx.media3.common.MediaItem.SubtitleConfiguration
-import androidx.media3.common.MimeTypes
 import androidx.media3.common.PlaybackException
 import androidx.media3.common.Player
 import androidx.media3.common.Tracks
@@ -34,19 +32,18 @@ import dagger.hilt.android.AndroidEntryPoint
 import dev.anilbeesetti.libs.ffcodecs.FfmpegRenderersFactory
 import dev.anilbeesetti.nextplayer.core.common.extensions.getFilenameFromUri
 import dev.anilbeesetti.nextplayer.core.common.extensions.getPath
-import dev.anilbeesetti.nextplayer.core.ui.R as coreUiR
 import dev.anilbeesetti.nextplayer.feature.player.databinding.ActivityPlayerBinding
 import dev.anilbeesetti.nextplayer.feature.player.dialogs.TrackSelectionFragment
 import dev.anilbeesetti.nextplayer.feature.player.extensions.isRendererAvailable
 import dev.anilbeesetti.nextplayer.feature.player.extensions.switchTrack
 import dev.anilbeesetti.nextplayer.feature.player.extensions.toMediaItem
-import dev.anilbeesetti.nextplayer.feature.player.extensions.toSubtitleConfiguration
 import dev.anilbeesetti.nextplayer.feature.player.extensions.toggleSystemBars
 import dev.anilbeesetti.nextplayer.feature.player.utils.PlayerGestureHelper
-import java.io.File
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
 import timber.log.Timber
+import java.io.File
+import dev.anilbeesetti.nextplayer.core.ui.R as coreUiR
 
 @SuppressLint("UnsafeOptInUsageError")
 @AndroidEntryPoint
@@ -230,45 +227,7 @@ class PlayerActivity : AppCompatActivity() {
                     val mediaItems = viewModel.currentPlayerItems.map { it.toMediaItem(this) }
                     player.setMediaItems(mediaItems, viewModel.currentPlayerItemIndex, C.TIME_UNSET)
                 } else {
-                    dataUri?.also {
-                        var defaultSub: Uri? = null
-                        val subsEnable = extras?.getParcelableArray(API_SUBS_ENABLE)
-                        if (!subsEnable.isNullOrEmpty()) {
-                            defaultSub = subsEnable[0] as Uri
-                        }
-                        val subtitleConfigurations = mutableListOf<SubtitleConfiguration>()
-                        if (extras?.containsKey(API_SUBS) == true) {
-                            val subs = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-                                extras?.getParcelableArray(API_SUBS, Uri::class.java)
-                            } else {
-                                extras?.getParcelableArray(API_SUBS)
-                            }
-                            val subsName = extras?.getStringArray(API_SUBS_NAME)
-                            Timber.d("Subs: $subs")
-                            if (!subs.isNullOrEmpty()) {
-                                for (i in subs.indices) {
-                                    var subName: String? = null
-                                    val subtitle = subs[i] as Uri
-                                    if (subsName != null && subsName.size > i) {
-                                        subName = subsName[i]
-                                    }
-                                    subtitleConfigurations.add(
-                                        subtitle.toSubtitleConfiguration(
-                                            context = this,
-                                            selected = subtitle == defaultSub,
-                                            name = subName
-                                        )
-                                    )
-                                }
-                            }
-                        }
-                        player.addMediaItem(
-                            MediaItem.Builder()
-                                .setUri(it)
-                                .setSubtitleConfigurations(subtitleConfigurations)
-                                .build()
-                        )
-                    }
+                    dataUri?.also { player.addMediaItem(it.toMediaItem(this, extras)) }
                     extras?.also {
                         if (it.containsKey(API_POSITION)) {
                             player.seekTo(it.getInt(API_POSITION).toLong())
@@ -401,14 +360,14 @@ class PlayerActivity : AppCompatActivity() {
     }
 
     companion object {
-        private const val API_TITLE = "title"
-        private const val API_POSITION = "position"
-        private const val API_DURATION = "duration"
-        private const val API_RETURN_RESULT = "return_result"
-        private const val API_END_BY = "end_by"
-        private const val API_SUBS = "subs"
-        private const val API_SUBS_ENABLE = "subs.enable"
-        private const val API_SUBS_NAME = "subs.name"
+        const val API_TITLE = "title"
+        const val API_POSITION = "position"
+        const val API_DURATION = "duration"
+        const val API_RETURN_RESULT = "return_result"
+        const val API_END_BY = "end_by"
+        const val API_SUBS = "subs"
+        const val API_SUBS_ENABLE = "subs.enable"
+        const val API_SUBS_NAME = "subs.name"
     }
 }
